# backend/app/auth/router.py
from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.security import OAuth2PasswordRequestForm
from sqlalchemy.orm import Session
from datetime import timedelta

from app.database import get_db
from app.models import User
from app.schemas import UserCreate, UserResponse, TokenResponse, UserLogin
from app.auth.jwt import (
    create_access_token,
    verify_password,
    get_password_hash,
    validate_password,
    ACCESS_TOKEN_EXPIRE_HOURS,
)
from app.auth.dependencies import get_current_user
from app.middleware.rate_limit import rate_limit
import logging

logger = logging.getLogger(__name__)
auth_router = APIRouter()


@auth_router.post(
    "/register",
    response_model=UserResponse,
    status_code=status.HTTP_201_CREATED,
    summary="Register new user",
    description="Create a new user account with username, email, and password",
    dependencies=[Depends(rate_limit("5/minute"))],
    responses={
        201: {
            "description": "User successfully created",
            "content": {
                "application/json": {
                    "example": {
                        "id": 1,
                        "username": "john_doe",
                        "email": "john.doe@example.com",
                        "is_admin": False,
                        "created_at": "2024-01-01T00:00:00",
                    }
                }
            },
        },
        400: {
            "description": "Bad request - Username/email already exists or password requirements not met",
            "content": {
                "application/json": {
                    "examples": {
                        "username_exists": {
                            "summary": "Username already taken",
                            "value": {"detail": "Username already registered", "status": 400},
                        },
                        "email_exists": {
                            "summary": "Email already registered",
                            "value": {"detail": "Email already registered", "status": 400},
                        },
                        "weak_password": {
                            "summary": "Password too weak",
                            "value": {
                                "detail": "Password does not meet security requirements",
                                "status": 400,
                            },
                        },
                    }
                }
            },
        },
    },
)
async def register(user_data: UserCreate, db: Session = Depends(get_db)):
    """
    Register a new user account.

    Password requirements:
    - Minimum 8 characters
    - Must contain uppercase and lowercase letters
    - Must contain at least one number
    - Must contain at least one special character

    Args:
        user_data: User registration information
        db: Database session

    Returns:
        UserResponse: Created user information (without password)

    Raises:
        HTTPException: If username/email already exists or password is invalid
    """

    # 사용자명 중복 검사
    if db.query(User).filter(User.username == user_data.username).first():
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST, detail="Username already registered"
        )

    # 이메일 중복 검사
    if db.query(User).filter(User.email == user_data.email).first():
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST, detail="Email already registered"
        )

    # 비밀번호 정책 검증
    if not validate_password(user_data.password):
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Password does not meet security requirements",
        )

    # 새 사용자 생성
    hashed_password = get_password_hash(user_data.password)
    db_user = User(
        username=user_data.username, email=user_data.email, password_hash=hashed_password
    )

    db.add(db_user)
    db.commit()
    db.refresh(db_user)

    return db_user


@auth_router.post(
    "/login",
    response_model=TokenResponse,
    summary="User login",
    description="Authenticate user and receive JWT token",
    dependencies=[Depends(rate_limit("10/minute"))],
    responses={
        200: {
            "description": "Successful authentication",
            "content": {
                "application/json": {
                    "example": {
                        "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                        "token_type": "bearer",
                        "user": {
                            "id": 1,
                            "username": "john_doe",
                            "email": "john.doe@example.com",
                            "is_admin": False,
                            "created_at": "2024-01-01T00:00:00",
                        },
                    }
                }
            },
        },
        401: {
            "description": "Invalid credentials",
            "content": {
                "application/json": {
                    "example": {"detail": "Incorrect username or password", "status": 401}
                }
            },
        },
    },
)
async def login(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(get_db)):
    """
    Authenticate user and generate JWT token.

    The token expires after 24 hours and must be included in subsequent requests
    as a Bearer token in the Authorization header.

    Args:
        form_data: OAuth2 form with username and password
        db: Database session

    Returns:
        TokenResponse: JWT access token and user information

    Raises:
        HTTPException: If credentials are invalid
    """

    # 사용자 인증
    user = db.query(User).filter(User.username == form_data.username).first()
    if not user or not verify_password(form_data.password, user.password_hash):
        logger.warning(f"Failed login attempt for username: {form_data.username}")
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Incorrect username or password",
            headers={"WWW-Authenticate": "Bearer"},
        )

    # JWT 토큰 생성
    access_token_expires = timedelta(hours=ACCESS_TOKEN_EXPIRE_HOURS)
    access_token = create_access_token(
        data={"sub": user.username}, expires_delta=access_token_expires
    )

    logger.info(f"User {user.username} logged in successfully")
    return {"access_token": access_token, "token_type": "bearer", "user": user}


@auth_router.post("/logout")
async def logout():
    """사용자 로그아웃 (클라이언트 측에서 토큰 삭제)"""
    return {"message": "Successfully logged out"}


@auth_router.get("/me", response_model=UserResponse)
async def get_current_user_info(current_user: User = Depends(get_current_user)):
    """현재 사용자 정보 조회"""
    return current_user