# backend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "LionRocket AI Chat API",
    "description": "AI-powered chat service using Claude 3 with JWT authentication",
    "version": "1.0.0",
    "contact": {
      "name": "LionRocket Support",
      "email": "support@lionrocket.com",
      "url": "https://lionrocket.com/support"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "https://api.lionrocket.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["authentication"],
        "summary": "Register new user",
        "operationId": "register_user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/UserCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["authentication"],
        "summary": "User login",
        "operationId": "login_user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/OAuth2PasswordRequestForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TokenResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/chats": {
      "get": {
        "tags": ["chats"],
        "summary": "Get user chats",
        "operationId": "get_user_chats",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ChatListResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["chats"],
        "summary": "Create new chat",
        "operationId": "create_chat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ChatResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/chats/{chat_id}/messages": {
      "post": {
        "tags": ["chats"],
        "summary": "Send message",
        "operationId": "send_message",
        "parameters": [
          {
            "$ref": "#/components/parameters/ChatIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MessageResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserCreate": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        }
      },
      "OAuth2PasswordRequestForm": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ChatCreate": {
        "type": "object",
        "required": ["character_id"],
        "properties": {
          "character_id": {
            "type": "integer"
          },
          "initial_message": {
            "type": "string",
            "maxLength": 200
          }
        }
      },
      "MessageCreate": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      }
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "ChatIdParam": {
        "name": "chat_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    },
    "responses": {
      "UserCreated": {
        "description": "User successfully created",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "id": {"type": "integer"},
                "username": {"type": "string"},
                "email": {"type": "string"},
                "is_admin": {"type": "boolean"},
                "created_at": {"type": "string", "format": "date-time"}
              }
            }
          }
        }
      },
      "TokenResponse": {
        "description": "Successful authentication",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {"type": "string"},
                "token_type": {"type": "string"},
                "user": {"type": "object"}
              }
            }
          }
        }
      },
      "ChatListResponse": {
        "description": "Paginated list of chats",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "items": {"type": "array"},
                "total": {"type": "integer"},
                "page": {"type": "integer"},
                "pages": {"type": "integer"},
                "limit": {"type": "integer"}
              }
            }
          }
        }
      },
      "ChatResponse": {
        "description": "Chat created successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "MessageResponse": {
        "description": "Message sent successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {"type": "string"},
                "status": {"type": "integer"}
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {"type": "string"},
                "status": {"type": "integer"}
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {"type": "string"},
                "status": {"type": "integer"}
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {"type": "array"},
                "status": {"type": "integer"}
              }
            }
          }
        }
      },
      "RateLimitExceeded": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {"type": "string"},
                "status": {"type": "integer"}
              }
            }
          }
        }
      }
    }
  }
}